+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.27.4...HEAD) - TBD
 
+### :bug: Fixed
+
+- Coverage phase probing methods of operations disabled via `[[operations]]`.
+
 ## [4.27.4](https://github.com/schemathesis/schemathesis/compare/v4.27.3...v4.27.4) - 2026-09-19
 
 ### :wrench: Changed