+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.27.4...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.27.5...HEAD) - TBD
+
+## [4.27.5](https://github.com/schemathesis/schemathesis/compare/v4.27.4...v4.27.5) - 2026-09-20
 
 ### :bug: Fixed